# docs/openapi.yaml
openapi: 3.0.3

info:
  title: Categories API
  version: 1.0.0
  description: Category management API for the application.

servers:
  - url: http://localhost:3000/api/v1
    description: Local development server

tags:
  - name: Categories
    description: Category management operations

paths:
  /categories:
    get:
      tags:
        - Categories
      summary: List categories
      operationId: listCategories
      parameters:
        - name: skip
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: take
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: includeChildren
          in: query
          required: false
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: Categories returned successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryListResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalServerError"

    post:
      tags:
        - Categories
      summary: Create a category
      operationId: createCategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryRequest"
            example:
              name: Shoes
              slug: shoes
              description: Footwear
      responses:
        "201":
          description: Category created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "409":
          description: Category slug already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"

  /categories/{id}:
    parameters:
      - $ref: "#/components/parameters/CategoryId"

    get:
      tags:
        - Categories
      summary: Get a category by ID
      operationId: getCategoryById
      responses:
        "200":
          description: Category returned successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalServerError"

    patch:
      tags:
        - Categories
      summary: Update a category
      operationId: updateCategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCategoryRequest"
      responses:
        "200":
          description: Category updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "409":
          description: Update conflicts with an existing category or hierarchy rule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"

    delete:
      tags:
        - Categories
      summary: Delete a category
      operationId: deleteCategory
      responses:
        "204":
          description: Category deleted successfully
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "409":
          description: Category has dependent children or products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"

  /categories/slug/{slug}:
    parameters:
      - name: slug
        in: path
        required: true
        description: URL-safe category slug
        schema:
          type: string
          minLength: 1
          maxLength: 180
          pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$"

    get:
      tags:
        - Categories
      summary: Get a category by slug
      operationId: getCategoryBySlug
      responses:
        "200":
          description: Category returned successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryResponse"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalServerError"

components:
  parameters:
    CategoryId:
      name: id
      in: path
      required: true
      description: SQL Server GUID generated by NEWSEQUENTIALID()
      schema:
        type: string
        pattern: "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$"
        example: "8f6c2e3a-4c1d-4b7a-9f25-2a6d7e8c9012"

  schemas:
    Category:
      type: object
      required:
        - id
        - name
        - slug
      properties:
        id:
          type: string
          pattern: "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$"
        name:
          type: string
          minLength: 1
          maxLength: 150
        slug:
          type: string
          minLength: 1
          maxLength: 180
          pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$"
        description:
          type: string
          nullable: true
          maxLength: 500
        parentId:
          type: string
          nullable: true
        parent:
          allOf:
            - $ref: "#/components/schemas/CategorySummary"
          nullable: true
        children:
          type: array
          items:
            $ref: "#/components/schemas/CategorySummary"

    CategorySummary:
      type: object
      required:
        - id
        - name
        - slug
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string

    CreateCategoryRequest:
      type: object
      required:
        - name
        - slug
      additionalProperties: false
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 150
        slug:
          type: string
          minLength: 1
          maxLength: 180
          pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$"
        description:
          type: string
          nullable: true
          maxLength: 500
        parentId:
          type: string
          nullable: true

    UpdateCategoryRequest:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 150
        slug:
          type: string
          minLength: 1
          maxLength: 180
          pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$"
        description:
          type: string
          nullable: true
          maxLength: 500
        parentId:
          type: string
          nullable: true

    CategoryResponse:
      type: object
      required:
        - data
      properties:
        data:
          $ref: "#/components/schemas/Category"

    CategoryListResponse:
      type: object
      required:
        - data
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Category"

    ErrorDetail:
      type: object
      properties:
        location:
          type: string
          example: params
        field:
          type: string
          example: id
        message:
          type: string
          example: id must be a valid GUID

    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
        details:
          type: array
          items:
            $ref: "#/components/schemas/ErrorDetail"

  responses:
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: Validation failed
            details:
              - location: params
                field: id
                message: id must be a valid GUID

    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: Category not found

    InternalServerError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: Internal server error